Revert of the commented 'created' puppet attribute

+++ manifests/phenotypedbapp.pp
@@ -78,7 +78,7 @@ define phenotypedb::phenotypedbapp (
         command => "/usr/bin/wget -O '${downloaded_war}' '${download_url}' \
                    && find '$deployment_dir' -name '*.war' -delete \
                    && mv '${downloaded_war}' '${deployed_war}'",
-        # creates => $deployed_war,
+        creates => $deployed_war,
         timeout => 1200,
     }
